Allow Bob.threshold_decrypt(...) to specify a decryption timeout - defaults to 15s.

+++ nucypher/network/decryption.py
@@ -27,7 +27,7 @@ class ThresholdDecryptionClient(ThresholdAccessControlClient):
         self,
         encrypted_requests: Dict[ChecksumAddress, EncryptedThresholdDecryptionRequest],
         threshold: int,
-        timeout: float = 15,
+        timeout: int = 15,
     ) -> Tuple[
         Dict[ChecksumAddress, EncryptedThresholdDecryptionResponse],
         Dict[ChecksumAddress, str],
